Add serial variant
